Update monitoring-metricbeat.asciidoc (#31795)
* Update monitoring-metricbeat.asciidoc

This is miss leading, cluster_uuid, is required to be able to see the monitoring data. It will only in kibana stack monitoring view if it is assigned a cluster_uuid to group the beats under.  So, requires a cluster_uuid. If not the data has empty cluster_uuid and will not be visible in stack monitoring area.  Index populate just not visibile in SM UI

* Update libbeat/docs/monitoring/monitoring-metricbeat.asciidoc

* Remove optional from description of cluster_id setting

Co-authored-by: DeDe Morton <dede.morton@elastic.co>